Fix broken world map
The world map broke after integrating the Naver Maps API, which mutates the `window.marker` variable on load; therefore, the `marker` object was lost. This code will create a deep copy of `marker` to `_marker`; however, it will lose info if the object contains non-JSON valid information (e.g. Date objects). This is not a permanent fix.